zvladol som to je to done je to krucipisek done uz len fakt mini chybicky opravit a je to hotovo god did.

+++ Project_Case/Skin.cs
@@ -10,11 +10,12 @@ namespace CS2_CaseOpening
         public string Rarity { get; set; }
         public string ImagePath { get; set; }
 
-        // Numeric float value between 0.0 and 1.0 (lower is generally better condition)
+       
         public double Float { get; set; }
 
         public string Wear { get; set; }
 
-        public int Price { get; set; }
+        
+        public double Price { get; set; }
     }
 }
